requirements.txt
async-timeout==4.0.2
attrs==22.2.0
autopep8==1.7.0
bcrypt==4.0.0
Brotli==1.0.9
click==8.1.3
Deprecated==1.2.13
dnspython==2.2.1
exceptiongroup==1.1.0
Flask==2.2.2
Flask-Compress==1.13
Flask-Cors==3.0.10
Flask-JWT==0.3.2
gevent==21.12.0
greenlet==1.1.3
gunicorn==20.1.0
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
mergedeep==1.3.4
mongoengine==0.24.2
packaging==21.3
pluggy==1.0.0
pycodestyle==2.9.1
PyJWT==1.4.2
pymongo==4.2.0
pyparsing==3.0.9
pytest==7.2.2
pytest-flask==1.2.0
PyYAML==6.0
redis==4.3.4
six==1.16.0
toml==0.10.2
tomli==2.0.1
Werkzeug==2.2.2
wrapt==1.14.1
zope.event==4.6
zope.interface==5.5.2
